refactor logic to create service client to ClientFactory
