+++ VedaConnect/VedaConnect.csproj
@@ -44,6 +44,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ClientFactory.cs" />
     <Compile Include="MessageInspector.cs" />
     <Compile Include="Models\Address.cs" />
     <Compile Include="Models\CurrentOrPrevious.cs" />
