+++ VedaConnect/VedaScoreApplyClient.cs
@@ -2,9 +2,6 @@
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
-using System.Net;
-using System.ServiceModel;
-using System.ServiceModel.Channels;
 using System.Threading.Tasks;
 using VedaConnect.VedaScoreApply;
 
@@ -18,33 +15,10 @@ namespace VedaConnect
     public class VedaScoreApplyClient : IVedaScoreApplyClient
     {
         private readonly VedaScoreApplyPortTypeClient _client;
-        private readonly MessageInspector _messageInspector;
 
         public VedaScoreApplyClient(string url, string username, string password)
         {
-            var encoding = new TextMessageEncodingBindingElement { MessageVersion = MessageVersion.Soap11WSAddressing10 };
-            var transport = new HttpsTransportBindingElement { AuthenticationScheme = AuthenticationSchemes.Basic, KeepAliveEnabled = true, MaxReceivedMessageSize = int.MaxValue };
-            var security = SecurityBindingElement.CreateUserNameOverTransportBindingElement();
-            security.IncludeTimestamp = false;
-
-            var binding = new CustomBinding();
-            binding.Elements.Add(encoding);
-            binding.Elements.Add(security);
-            binding.Elements.Add(transport);
-
-            var endpointAddress = new EndpointAddress(url);
-
-            _messageInspector = new MessageInspector();
-            _client = new VedaScoreApplyPortTypeClient(binding, endpointAddress);
-            _client.Endpoint.Behaviors.Add(_messageInspector);
-
-            if (_client.ClientCredentials == null)
-            {
-                throw new Exception("Client credentials not available because the web service endpoint hasn't been configured properly.");
-            }
-
-            _client.ClientCredentials.UserName.UserName = username;
-            _client.ClientCredentials.UserName.Password = password;
+            _client = ClientFactory.Create<VedaScoreApplyPortTypeClient, VedaScoreApplyPortType>(url, username, password);
         }
 
         public async Task<SubmitEnquiryResult> SubmitEnquiryAsync(Enquiry enquiry)
